Move slippage service to its own file
