+++ libs/services/src/index.ts
@@ -1,2 +1,4 @@
 export * from './SlippageService/SlippageService';
+export * from './SlippageService/SlippageServiceMain';
+
 export * from './UsdService/UsdService';
